+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+ - 2017-02-16
+   - Added support for rendering markdown partials
 ## [Release 1.1.0](https://github.com/QutBioacoustics/baw-server/releases/tag/1.1.0)
 
 Our major new feature in this release is support for user based analysis-jobs!